requirements.txt
algoliasearch==2.4.0
certifi==2020.6.20
chardet==3.0.4
deprecated==1.2.10
emoji==0.6.0
idna==2.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygithub==1.53
pyjwt==1.7.1
python-gitlab==2.5.0
python-ldap==3.3.1
requests==2.24.0
urllib3==1.25.10
wrapt==1.12.1
